View pra Update de SistemaUsuario que envia pro form de alteracao do usuario selecionado e outras alteracoes

+++ src/main/java/br/com/fzlbpms/model/common/Pessoa.java
@@ -4,15 +4,12 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToOne;
-
-import br.com.fzlbpms.model.sistema.SistemaUsuario;
 
 @Entity
 public class Pessoa {
 
 	@Id @GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String nome;
 	private String rg;
 	private String cpf;
@@ -21,11 +18,11 @@ public class Pessoa {
 	public Pessoa() {		
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
